Add files via upload
I can't insert the late_fee stuff into the script because the date_borrowed and date_return data is always changing and needs to be inserted from the query because its not hard data. Its probably possible to do it but its overly complicated and not something we've gone over.